+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.10.0]
+### Added
+- King's Fall loot table (missing weapons due to classified items)
+- New command: `/roll-finder` uses curated weapon rolls to provide recommended rolls for PvE/PvP
+
 ## [6.9.0]
 ### Added
 - New craftable weapons.